refactor: folder 페이지의 내용을 FolderContent로 분리
